+++ src/fortran_routing/mc_pylink_v00/MC_singleSeg_singleTS/varSingleSegStime_f2py_2clean.f90
@@ -2,15 +2,15 @@ module var
     implicit none
     save
 
-    real :: dt, dx, qup, quc, qdp, qdc, ql
-    real :: Bw, Tw, TwCC, nCC, Cs, So, n, z, vel, depth
-    real :: velp_chk, depthp_chk
-    real :: bfd, WPC, AREAC, C1, C2, C3, C4
+    real*8 :: dt, dx, qup, quc, qdp, qdc, ql
+    real*8 :: Bw, Tw, TwCC, nCC, Cs, So, n, z, vel, depth
+    real*8 :: velp_chk, depthp_chk
+    real*8 :: bfd, WPC, AREAC, C1, C2, C3, C4
     integer :: ntim
     integer :: ncomp0, ncomp, iseg, uslinkflag
-    real :: Qus_prev
-    real,allocatable,dimension(:,:) :: vela, deptha
-    real,allocatable,dimension(:,:) :: Qd
+    real*8 :: Qus_prev
+    real*8,allocatable,dimension(:,:) :: vela, deptha
+    real*8,allocatable,dimension(:,:) :: Qd
 
 
 end module var
